+++ packages/api-graphql/src/internals/operations/subscription.ts
@@ -2,19 +2,20 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { map } from 'rxjs';
-import { V6Client, GraphqlSubscriptionResult } from '../../types';
-import {
-	initializeModel,
-	generateGraphQLDocument,
-	buildGraphQLVariables,
-	authModeParams,
-	ModelOperation,
-	getCustomHeaders,
-} from '../APIClient';
 import {
 	ModelIntrospectionSchema,
 	SchemaModel,
 } from '@aws-amplify/core/internals/utils';
+
+import { GraphqlSubscriptionResult, V6Client } from '../../types';
+import {
+	ModelOperation,
+	authModeParams,
+	buildGraphQLVariables,
+	generateGraphQLDocument,
+	getCustomHeaders,
+	initializeModel,
+} from '../APIClient';
 
 export function subscriptionFactory(
 	client: any,
@@ -29,7 +30,7 @@ export function subscriptionFactory(
 			modelIntrospection,
 			name,
 			operation,
-			args
+			args,
 		);
 
 		const variables = buildGraphQLVariables(
@@ -64,6 +65,7 @@ export function subscriptionFactory(
 					auth.authMode,
 					auth.authToken,
 				);
+
 				return initialized;
 			}),
 		);