+++ packages/geo/src/types/Geo.ts
@@ -53,20 +53,20 @@ export type SearchByTextOptions =
 	| SearchByTextOptionsWithSearchAreaConstraints;
 
 // Options object for searchByCoordinates
-export type SearchByCoordinatesOptions = {
+export interface SearchByCoordinatesOptions {
 	maxResults?: number;
 	searchIndexName?: string;
 	providerName?: string;
-};
+}
 
-export type searchByPlaceIdOptions = {
+export interface searchByPlaceIdOptions {
 	searchIndexName?: string;
-};
+}
 
 // Geometry object for Place points
-export type PlaceGeometry = {
+export interface PlaceGeometry {
 	point: Coordinates;
-};
+}
 
 // Place object with locality information
 export interface Place {
@@ -88,38 +88,38 @@ export type LinearRing = Coordinates[];
 export type GeofencePolygon = LinearRing[];
 
 // Geometry object for Polygon
-export type PolygonGeometry = {
+export interface PolygonGeometry {
 	polygon: GeofencePolygon;
-};
+}
 
 export type GeofenceId = string;
 
 // Geofence object used as input for saveGeofences
-export type GeofenceInput = {
+export interface GeofenceInput {
 	geofenceId: GeofenceId;
 	geometry: PolygonGeometry;
-};
+}
 
 // Options object for saveGeofences
-export type GeofenceOptions = {
+export interface GeofenceOptions {
 	providerName?: string;
-};
+}
 
 // Error type for errors related to Geofence API calls
-export type GeofenceError = {
+export interface GeofenceError {
 	error: {
 		code: string;
 		message: string;
 	};
 	geofenceId: GeofenceId;
-};
+}
 
 // Base geofence object
-type GeofenceBase = {
+interface GeofenceBase {
 	geofenceId: GeofenceId;
 	createTime?: Date;
 	updateTime?: Date;
-};
+}
 
 // Results object for getGeofence
 export type Geofence = GeofenceBase & {
@@ -127,10 +127,10 @@ export type Geofence = GeofenceBase & {
 };
 
 // Results object for saveGeofences
-export type SaveGeofencesResults = {
+export interface SaveGeofencesResults {
 	successes: GeofenceBase[];
 	errors: GeofenceError[];
-};
+}
 
 // Options object for listGeofence
 export type ListGeofenceOptions = GeofenceOptions & {
@@ -138,21 +138,21 @@ export type ListGeofenceOptions = GeofenceOptions & {
 };
 
 // Results options for listGeofence
-export type ListGeofenceResults = {
+export interface ListGeofenceResults {
 	entries: Geofence[];
 	nextToken: string | undefined;
-};
+}
 
 // Results object for deleteGeofence
-export type DeleteGeofencesResults = {
+export interface DeleteGeofencesResults {
 	successes: GeofenceId[];
 	errors: GeofenceError[];
-};
+}
 
 // Return type for searchForSuggestions
 export type SearchForSuggestionsResults = SearchForSuggestionsResult[];
 
-export type SearchForSuggestionsResult = {
+export interface SearchForSuggestionsResult {
 	text: string;
 	placeId?: string;
-};
+}