+++ packages/interactions/src/lex-v1/apis/onComplete.ts
@@ -5,8 +5,8 @@ import { OnCompleteInput } from '../types';
 import { resolveBotConfig } from '../utils';
 import { lexProvider } from '../AWSLexProvider';
 import {
-	assertValidationError,
 	InteractionsValidationErrorCode,
+	assertValidationError,
 } from '../../errors';
 
 export const onComplete = (input: OnCompleteInput): void => {