+++ packages/interactions/src/lex-v2/AWSLexV2Provider.ts
@@ -1,11 +1,6 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 import {
-	InteractionsOnCompleteCallback,
-	InteractionsMessage,
-	InteractionsResponse,
-} from '../types/Interactions';
-import {
 	IntentState,
 	LexRuntimeV2Client,
 	RecognizeTextCommand,
@@ -17,8 +12,15 @@ import {
 } from '@aws-sdk/client-lex-runtime-v2';
 import { getAmplifyUserAgentObject } from '@aws-amplify/core/internals/utils';
 import { ConsoleLogger, fetchAuthSession } from '@aws-amplify/core';
-import { convert, unGzipBase64AsJson } from '../utils';
 import { v4 as uuid } from 'uuid';
+
+import { convert, unGzipBase64AsJson } from '../utils';
+import {
+	InteractionsMessage,
+	InteractionsOnCompleteCallback,
+	InteractionsResponse,
+} from '../types/Interactions';
+
 import { AWSLexV2ProviderOption } from './types';
 
 const logger = new ConsoleLogger('AWSLexV2Provider');
@@ -43,18 +45,19 @@ type AWSLexV2ProviderSendResponse =
 	| RecognizeTextCommandOutput
 	| RecognizeUtteranceCommandOutputFormatted;
 
-type lexV2BaseReqParams = {
+interface lexV2BaseReqParams {
 	botId: string;
 	botAliasId: string;
 	localeId: string;
 	sessionId: string;
-};
+}
 
 class AWSLexV2Provider {
 	private readonly _botsCompleteCallback: Record<
 		string,
 		InteractionsOnCompleteCallback
 	> = {};
+
 	private defaultSessionId: string = uuid();
 
 	/**
@@ -73,7 +76,7 @@ class AWSLexV2Provider {
 		try {
 			session = await fetchAuthSession();
 		} catch (error) {
-			return Promise.reject('No credentials');
+			return Promise.reject(new Error('No credentials'));
 		}
 
 		const { region, aliasId, localeId, botId } = botConfig;
@@ -108,6 +111,7 @@ class AWSLexV2Provider {
 				client,
 			);
 		}
+
 		return response;
 	}
 
@@ -193,10 +197,11 @@ class AWSLexV2Provider {
 
 		try {
 			const recognizeTextCommand = new RecognizeTextCommand(params);
-			const data = await client.send(recognizeTextCommand);
+			const resultData = await client.send(recognizeTextCommand);
 
-			this._reportBotStatus(data, botConfig);
-			return data;
+			this._reportBotStatus(resultData, botConfig);
+
+			return resultData;
 		} catch (err) {
 			return Promise.reject(err);
 		}
@@ -223,7 +228,7 @@ class AWSLexV2Provider {
 		// prepare params
 		if (messageType === 'voice') {
 			if (typeof content !== 'object') {
-				return Promise.reject('invalid content type');
+				return Promise.reject(new Error('invalid content type'));
 			}
 
 			const inputStream =
@@ -237,7 +242,7 @@ class AWSLexV2Provider {
 		} else {
 			// text input
 			if (typeof content !== 'string')
-				return Promise.reject('invalid content type');
+				return Promise.reject(new Error('invalid content type'));
 
 			params = {
 				...baseParams,
@@ -249,10 +254,11 @@ class AWSLexV2Provider {
 		// make API call to lex
 		try {
 			const recognizeUtteranceCommand = new RecognizeUtteranceCommand(params);
-			const data = await client.send(recognizeUtteranceCommand);
+			const resultData = await client.send(recognizeUtteranceCommand);
 
-			const response = await this._formatUtteranceCommandOutput(data);
+			const response = await this._formatUtteranceCommandOutput(resultData);
 			this._reportBotStatus(response, botConfig);
+
 			return response;
 		} catch (err) {
 			return Promise.reject(err);