+++ packages/predictions/src/providers/AmazonAIIdentifyPredictionsProvider.ts
@@ -26,6 +26,7 @@ import {
 	DetectDocumentTextCommandInput,
 	TextractClient,
 } from '@aws-sdk/client-textract';
+
 import { PredictionsValidationErrorCode } from '../errors/types/validation';
 import { assertValidationError } from '../errors/utils/assertValidationError';
 import {
@@ -50,12 +51,8 @@ import {
 	isStorageSource,
 	isValidIdentifyInput,
 } from '../types';
-import {
-	BlockList,
-	Document,
-	Image,
-	TextDetectionList,
-} from '../types/AWSTypes';
+import { BlockList, Image, TextDetectionList } from '../types/AWSTypes';
+
 import {
 	categorizeRekognitionBlocks,
 	categorizeTextractBlocks,
@@ -84,12 +81,15 @@ export class AmazonAIIdentifyPredictionsProvider {
 
 		if (isIdentifyTextInput(input)) {
 			logger.debug('identifyText');
+
 			return this.identifyText(input);
 		} else if (isIdentifyLabelsInput(input)) {
 			logger.debug('identifyLabels');
+
 			return this.identifyLabels(input);
 		} else {
 			logger.debug('identifyEntities');
+
 			return this.identifyEntities(input);
 		}
 	}
@@ -102,7 +102,7 @@ export class AmazonAIIdentifyPredictionsProvider {
 	 * @return {Promise<Image>} - Promise resolving to the converted source object.
 	 */
 	private configureSource(source: IdentifySource): Promise<Image> {
-		return new Promise((res, rej) => {
+		return new Promise((resolve, reject) => {
 			if (isStorageSource(source)) {
 				const storageConfig = {
 					accessLevel: source.level,
@@ -114,37 +114,44 @@ export class AmazonAIIdentifyPredictionsProvider {
 						const parser =
 							/https:\/\/([a-zA-Z0-9%\-_.]+)\.s3\.[A-Za-z0-9%\-._~]+\/([a-zA-Z0-9%\-._~/]+)\?/;
 						const parsedURL = value.url.toString().match(parser) ?? '';
-						if (parsedURL.length < 3) rej('Invalid S3 key was given.');
-						res({
+						if (parsedURL.length < 3)
+							reject(new Error('Invalid S3 key was given.'));
+						resolve({
 							S3Object: {
 								Bucket: parsedURL[1],
 								Name: decodeURIComponent(parsedURL[2]),
 							},
 						});
 					})
-					.catch(err => rej(err));
+					.catch(err => {
+						reject(err);
+					});
 			} else if (isFileSource(source)) {
 				blobToArrayBuffer(source.file)
 					.then(buffer => {
-						res({ Bytes: new Uint8Array(buffer) });
+						resolve({ Bytes: new Uint8Array(buffer) });
 					})
-					.catch(err => rej(err));
+					.catch(err => {
+						reject(err);
+					});
 			} else if (isIdentifyBytesSource(source)) {
-				const bytes = source.bytes;
+				const { bytes } = source;
 				if (bytes instanceof Blob) {
 					blobToArrayBuffer(bytes)
 						.then(buffer => {
-							res({ Bytes: new Uint8Array(buffer) });
+							resolve({ Bytes: new Uint8Array(buffer) });
 						})
-						.catch(err => rej(err));
+						.catch(err => {
+							reject(err);
+						});
 				}
 				if (bytes instanceof ArrayBuffer || bytes instanceof Buffer) {
-					res({ Bytes: new Uint8Array(bytes) } as Image);
+					resolve({ Bytes: new Uint8Array(bytes) } as Image);
 				}
 				// everything else can be directly passed to Rekognition / Textract.
-				res({ Bytes: bytes } as Image);
+				resolve({ Bytes: bytes } as Image);
 			} else {
-				rej('Input source is not configured correctly.');
+				reject(new Error('Input source is not configured correctly.'));
 			}
 		});
 	}
@@ -179,9 +186,8 @@ export class AmazonAIIdentifyPredictionsProvider {
 			credentials,
 			customUserAgent: _getPredictionsIdentifyAmplifyUserAgent(),
 		});
-		let inputDocument: Document;
 
-		inputDocument = await this.configureSource(input.text?.source);
+		const inputDocument = await this.configureSource(input.text?.source);
 
 		// get default value if format isn't specified in the input.
 		const format = input.text?.format ?? configFormat;
@@ -237,6 +243,7 @@ export class AmazonAIIdentifyPredictionsProvider {
 
 			const analyzeDocumentCommand = new AnalyzeDocumentCommand(param);
 			const { Blocks } = await this.textractClient.send(analyzeDocumentCommand);
+
 			return categorizeTextractBlocks(Blocks as BlockList);
 		}
 	}
@@ -285,14 +292,15 @@ export class AmazonAIIdentifyPredictionsProvider {
 			data.forEach(val => {
 				identifyResult = { ...identifyResult, ...val };
 			});
+
 			return identifyResult;
 		});
 	}
 
 	/**
 	 * Calls Rekognition.detectLabels and organizes the returned data.
-	 * @param {DetectLabelsInput} param - parameter to be passed onto Rekognition
-	 * @return {Promise<IdentifyLabelsOutput>} - Promise resolving to organized detectLabels response.
+	 * @param param - parameters as {@link DetectLabelsCommandInput} to be passed onto Rekognition
+	 * @return a promise resolving to organized detectLabels response as {@link IdentifyLabelsOutput}.
 	 */
 	private async detectLabels(
 		param: DetectLabelsCommandInput,
@@ -306,6 +314,7 @@ export class AmazonAIIdentifyPredictionsProvider {
 					instance =>
 						makeCamelCase(instance.BoundingBox) as BoundingBox | undefined,
 				) || [];
+
 			return {
 				name: label.Name,
 				boundingBoxes: boxes,
@@ -315,13 +324,14 @@ export class AmazonAIIdentifyPredictionsProvider {
 				},
 			};
 		});
+
 		return { labels: detectLabelData };
 	}
 
 	/**
 	 * Calls Rekognition.detectModerationLabels and organizes the returned data.
-	 * @param {Rekognition.DetectLabelsRequest} param - Parameter to be passed onto Rekognition
-	 * @return {Promise<IdentifyLabelsOutput>} - Promise resolving to organized detectModerationLabels response.
+	 * @param param parameter to be passed onto Rekognition as {@link DetectModerationLabelsCommandInput}
+	 * @return a promise resolving to organized detectModerationLabels response as {@link IdentifyLabelsOutput}.
 	 */
 	private async detectModerationLabels(
 		param: DetectModerationLabelsCommandInput,
@@ -342,8 +352,8 @@ export class AmazonAIIdentifyPredictionsProvider {
 	/**
 	 * Identify faces within an image that is provided as input, and match faces from a collection
 	 * or identify celebrities.
-	 * @param {IdentifyEntityInput} input - object containing the source image and face match options.
-	 * @return {Promise<IdentifyEntityOutput>} Promise resolving to identify results.
+	 * @param input - object of {@link IdentifyEntitiesInput} containing the source image and face match options.
+	 * @return a promise resolving to identify results as {@link IdentifyEntitiesOutput}.
 	 */
 	protected async identifyEntities(
 		input: IdentifyEntitiesInput,
@@ -401,6 +411,7 @@ export class AmazonAIIdentifyPredictionsProvider {
 							},
 						}) as IdentifyEntity,
 				) ?? [];
+
 			return { entities: faces };
 		} else if (
 			isIdentifyFromCollection(input.entities) &&
@@ -425,6 +436,7 @@ export class AmazonAIIdentifyPredictionsProvider {
 					const externalImageId = match.Face?.ExternalImageId
 						? this.decodeExternalImageId(match.Face.ExternalImageId)
 						: undefined;
+
 					return {
 						boundingBox: makeCamelCase(match.Face?.BoundingBox),
 						metadata: {
@@ -433,6 +445,7 @@ export class AmazonAIIdentifyPredictionsProvider {
 						},
 					} as IdentifyEntity;
 				}) ?? [];
+
 			return { entities: faces };
 		} else {
 			const detectFacesCommand = new DetectFacesCommand(param);
@@ -458,6 +471,7 @@ export class AmazonAIIdentifyPredictionsProvider {
 					faceAttributes.emotions = detail.Emotions?.map(
 						emotion => emotion.Type,
 					);
+
 					return {
 						boundingBox: makeCamelCase(detail.BoundingBox),
 						landmarks: makeCamelCaseArray(detail.Landmarks),
@@ -469,6 +483,7 @@ export class AmazonAIIdentifyPredictionsProvider {
 						},
 					} as IdentifyEntity;
 				}) ?? [];
+
 			return { entities: faces };
 		}
 	}