+++ packages/predictions/src/providers/index.ts
@@ -3,6 +3,7 @@
 import { AmazonAIConvertPredictionsProvider } from './AmazonAIConvertPredictionsProvider';
 import { AmazonAIIdentifyPredictionsProvider } from './AmazonAIIdentifyPredictionsProvider';
 import { AmazonAIInterpretPredictionsProvider } from './AmazonAIInterpretPredictionsProvider';
+
 export {
 	AmazonAIConvertPredictionsProvider,
 	AmazonAIIdentifyPredictionsProvider,