+++ packages/api/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.0.28](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@6.0.27...@aws-amplify/api@6.0.28) (2024-04-09)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+## 6.0.27 (2024-04-02)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+## 6.0.26 (2024-04-01)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+## 6.0.25 (2024-03-30)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+## 6.0.24 (2024-03-29)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+## 6.0.23 (2024-03-25)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+## 6.0.22 (2024-03-25)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+## 6.0.21 (2024-03-19)
+
+**Note:** Version bump only for package @aws-amplify/api
+
 ## 6.0.20 (2024-03-11)
 
 **Note:** Version bump only for package @aws-amplify/api